LPR/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
Flask==1.1.2
numpy==1.20.0
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==7.0.0
pytesseract==0.3.7
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0

